+++ src/test/java/org/mule/modules/ftpclient/sftp/SftpPollTest.java
@@ -124,7 +124,7 @@ public class SftpPollTest extends AbstractSftpClientTest {
 
     private void checkFile(MuleMessage message, byte[] content, String originalFilename, String filename, long length) {
         assertEquals(4, message.getPropertyNames(PropertyScope.INBOUND).size());
-        assertEquals(length, message.getInboundProperty("fileSize"));
+        assertEquals(length, (long)message.getInboundProperty("fileSize"));
         assertEquals(originalFilename, message.getInboundProperty("originalFilename"));
         assertEquals(filename, message.getInboundProperty("filename"));
         assertNotNull(message.getInboundProperty("timestamp"));
